+++ gmailconnector/__init__.py
@@ -5,4 +5,4 @@ from gmailconnector.read_email import ReadEmail  # noqa: F401
 from gmailconnector.send_email import SendEmail  # noqa: F401
 from gmailconnector.send_sms import SendSMS  # noqa: F401
 
-version = "0.6.9"
+version = "0.7.0"